+++ src/components/utilities/Defense/Overview/Table/DefenseOverviewTableRow.vue
@@ -13,7 +13,13 @@
     <td>{{ label }}</td>
 
     <td v-for="tableHeader in tableHeaders" v-show="tableHeader.visible" :key="tableHeader.key">
-      {{ statForHeader(tableHeader) }}
+      <template v-if="tableHeader.key !== 'totalDps'">{{ statForHeader(tableHeader) }}</template>
+      
+      <HtmlTooltip class="html-tooltip--critical-tooltip" v-else>
+        <template #trigger><span class="tooltip__text">{{ statForHeader(tableHeader) }}</span></template>
+        <span class="html-tooltip__text--non-critical">Non-crit: {{ Math.round(defenseStats.attackDamage ? defenseStats.attackDamage : defenseStats.totalDps).toLocaleString('en-US') }}</span> <br />
+        <span class="html-tooltip__text--critical">Crit: {{ Math.round(defenseStats.attackDamage ? defenseStats.attackDamage * (1 + defenseStats.criticalDamage / 100) : defenseStats.totalDps).toLocaleString('en-US') }}</span>
+      </HtmlTooltip>
 
       <HtmlTooltip v-if="isBuffDefense && (tableHeader.key === 'criticalDamage' || tableHeader.key === 'tooltipDps')">
         This is the bonus applied to other defenses