+++ src/defense_stats/BlazingPhoenixStat.ts
@@ -8,20 +8,17 @@ import DamageType from "@/enums/DamageType";
 
 export default class BlazingPhoenixStat implements DefenseStatInterface<string> {
     private readonly blazingPhoenixDamage: number
-    private readonly criticalMultiplier: number;
-    private readonly calculationConditions: CalculationConditionsInterface;
 
     constructor(
         defense: UserDataStoreDefenseInterface,
-        calculationConditions: CalculationConditionsInterface,
+        private readonly calculationConditions: CalculationConditionsInterface,
         defensePowerAdditives: ComputedRef<number>,
-        criticalMultiplier: ComputedRef<number>,
+        private readonly criticalMultiplier: ComputedRef<number>,
+        private readonly criticalDamage: ComputedRef<number>,
         shard: ShardInterface,
         modsShardsMultiplier: number,
     ) {
-        this.criticalMultiplier = criticalMultiplier.value
-        this.calculationConditions = calculationConditions
-        this.blazingPhoenixDamage = this.calculateBlazingPhoenixDamage(defense, defensePowerAdditives, shard) * modsShardsMultiplier * this.criticalMultiplier
+        this.blazingPhoenixDamage = this.calculateBlazingPhoenixDamage(defense, defensePowerAdditives, shard) * modsShardsMultiplier * this.criticalMultiplier.value
     }
 
     get label(): string 
@@ -37,7 +34,17 @@ export default class BlazingPhoenixStat implements DefenseStatInterface<string> 
     // Blazing Phoenix can only proc once every 10 seconds, so we divide the damage by 10 to get the damage per second
     get dps(): number
     {
-        return this.blazingPhoenixDamage / 10
+        return this.blazingPhoenixDamage / 10 * this.criticalMultiplier.value
+    }
+
+    get attackDamage(): number
+    {
+        return this.blazingPhoenixDamage / this.criticalMultiplier.value
+    }
+
+    get critDamage(): number
+    {
+        return this.blazingPhoenixDamage / this.criticalMultiplier.value * (1 + this.criticalDamage.value / 100)
     }
 
     private calculateBlazingPhoenixDamage(defense: UserDataStoreDefenseInterface, defensePowerAdditives: ComputedRef<number>, shard: ShardInterface): number {