+++ src/enums/EnumCollection.ts
@@ -1,4 +1,4 @@
-import EnumDefinition from "./EnumDefinition";
+import type EnumDefinition from "./EnumDefinition";
 
 export default class EnumCollection<TEnum extends EnumDefinition> {
     constructor(
@@ -11,5 +11,9 @@ export default class EnumCollection<TEnum extends EnumDefinition> {
 
     public contains(expectedEnum: TEnum): boolean {
         return this.enums.some((currentEnum: TEnum) => currentEnum.equals(expectedEnum));
+    }
+    
+    get length(): number {
+        return this.enums.length;
     }
 }